services/web/requirements.txt
flask==1.1.1
flask-sqlalchemy===2.4.1
flask-migrate==2.5.2
flask-login
flask-wtf
flask-ckeditor
flask-bcrypt
flask-mail
flask-moment
flask-babel
psycopg2-binary==2.8.4
gunicorn==20.0.4
email-validator
pyjwt